requirements.txt
tk==0.1.0
tkcalendar==1.6.1
python-amazon-sp-api==0.17.2
python-dateutil==2.8.2
XlsxWriter==3.0.3
pytz==2022.2.1
pandas==1.5.0
psycopg2-binary==2.9.4